refactor(ui): modularize UIOrchestration into dedicated folder; keep stable shim and global attach; add index.js + shared module layout
