+++ ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikaa/core/bridge/ui_orchestration/index.js
@@ -1,0 +1,131 @@
+import { attachGlobal } from "../../../utils/global.js";
+
+import { requestDraw } from "./draw.js";
+import { updateStats } from "./stats.js";
+import { renderInventory, showInventory, hideInventory, isInventoryOpen } from "./inventory.js";
+import { showLoot, hideLoot, isLootOpen } from "./loot.js";
+import {
+  showGameOver,
+  hideGameOver,
+  showGod,
+  hideGod,
+  isGodOpen,
+  showHelp,
+  hideHelp,
+  isHelpOpen,
+  showCharacter,
+  hideCharacter,
+  isCharacterOpen,
+  showShop,
+  hideShop,
+  isShopOpen,
+  buyShopIndex,
+  showSmoke,
+  hideSmoke,
+  isSmokeOpen,
+  showSleep,
+  hideSleep,
+  isSleepOpen,
+  animateSleep,
+} from "./panels.js";
+import { showConfirm, cancelConfirm, isConfirmOpen, isAnyModalOpen } from "./confirm.js";
+import { showFollower, hideFollower, isFollowerOpen } from "./follower.js";
+import { showQuestBoard, hideQuestBoard, isQuestBoardOpen } from "./quest_board.js";
+import { showLockpick, hideLockpick, isLockpickOpen } from "./lockpick.js";
+
+export {
+  requestDraw,
+  updateStats,
+  renderInventory,
+  showInventory,
+  hideInventory,
+  isInventoryOpen,
+  showLoot,
+  hideLoot,
+  isLootOpen,
+  showGameOver,
+  hideGameOver,
+  showGod,
+  hideGod,
+  isGodOpen,
+  showHelp,
+  hideHelp,
+  isHelpOpen,
+  showCharacter,
+  hideCharacter,
+  isCharacterOpen,
+  showShop,
+  hideShop,
+  isShopOpen,
+  buyShopIndex,
+  showSmoke,
+  hideSmoke,
+  isSmokeOpen,
+  showSleep,
+  hideSleep,
+  isSleepOpen,
+  animateSleep,
+  showConfirm,
+  cancelConfirm,
+  isConfirmOpen,
+  isAnyModalOpen,
+  showFollower,
+  hideFollower,
+  isFollowerOpen,
+  showQuestBoard,
+  hideQuestBoard,
+  isQuestBoardOpen,
+  showLockpick,
+  hideLockpick,
+  isLockpickOpen,
+};
+
+attachGlobal("UIOrchestration", {
+  requestDraw,
+  updateStats,
+  renderInventory,
+  showInventory,
+  hideInventory,
+  isInventoryOpen,
+  showLoot,
+  hideLoot,
+  isLootOpen,
+  showGameOver,
+  hideGameOver,
+  showGod,
+  hideGod,
+  isGodOpen,
+  showHelp,
+  hideHelp,
+  isHelpOpen,
+  showCharacter,
+  hideCharacter,
+  isCharacterOpen,
+  showShop,
+  hideShop,
+  isShopOpen,
+  buyShopIndex,
+  showSmoke,
+  hideSmoke,
+  isSmokeOpen,
+  showSleep,
+  hideSleep,
+  isSleepOpen,
+  animateSleep,
+  showConfirm,
+  cancelConfirm,
+  isConfirmOpen,
+  isAnyModalOpen,
+  // Follower inspect panel
+  showFollower,
+  hideFollower,
+  isFollowerOpen,
+  // Quest Board panel
+  showQuestBoard,
+  hideQuestBoard,
+  isQuestBoardOpen,
+  // Lockpicking modal
+  showLockpick,
+  hideLockpick,
+  isLockpickOpen,
+});
