+++ ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikaa/core/bridge/ui_orchestration/stats.js
@@ -1,0 +1,8 @@
+import { U } from "./shared.js";
+
+export function updateStats(ctx) {
+  const u = U(ctx);
+  if (u && typeof u.updateStats === "function") {
+    u.updateStats(ctx);
+  }
+}
