+++ ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikaa/docs/index.html
@@ -122,6 +122,7 @@
         { title: 'Engineering: Refactor & Modularization Plan', path: 'docs/refactor_modularization_plan.md' },
         { title: 'Engineering: GMBridge Modularization (Slice 1)', path: 'docs/gm_bridge_modularization.md' },
         { title: 'Engineering: Next Steps', path: 'docs/next_steps.md' },
+        { title: 'Engineering: UIOrchestration Modularization (Slice A)', path: 'docs/ui_orchestration_modularization.md' },
 
         { title: 'Data Docs: Index', path: 'data/docs/README.md' },
         { title: 'Data Docs: Palette Schema', path: 'data/docs/palette_schema.md' },
@@ -160,6 +161,8 @@
         'smoketest.md': 'Smoke test documentation and scenario notes.',
         'docs/refactor_modularization_plan.md': 'Engineering plan for modularizing/refactoring the runtime safely (guardrails + workstreams).',
         'docs/gm_bridge_modularization.md': 'Concrete file-split plan for core/bridge/gm_bridge.js (Slice 1).',
+        'docs/next_steps.md': 'What to do next: QA gates, invariants, and the next modularization slice to pick.',
+        'docs/ui_orchestration_modularization.md': 'UIOrchestration file split: new module layout + invariants.',
 
         'data/docs/README.md': 'Index of data documentation and palette guides.',
         'data/docs/palette_schema.md': 'Schema for palette JSON: keys, sections, and constraints.',
