+++ ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikoa/.gitignore
@@ -1,0 +1,2 @@
+# Ignore accidental typo folder created by editor/tooling
+/core/
